// src/payload-types.ts
export interface Media {
  id: number;
  /**
   * If left empty, this will be auto-populated from the filename.
   */
  alt?: string | null;
  /**
   * Categorizing media helps in organizing the library and Cloudinary folders.
   */
  category?:
    | (
        | 'ypsl'
        | 'executive-committees'
        | 'standing-committees'
        | 'ai-driven-sri-lanka'
        | 'sl-inspire'
        | 'lets-talk'
        | 'insl'
        | 'studpro'
        | 'y2npro'
        | 'merch-payslips'
        | 'others'
        | 'merchants'
        | 'ieee-projects'
        | 'events'
      )
    | null;
  cloudinaryUrl?: string | null;
  cloudinaryPublicId?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "awards".
 */
export interface Award {
  id: number;
  awardName: string;
  awardImage?: (number | null) | Media;
  awardCategory?: string | null;
  year?: string | null;
  winnerName: string;
  ouName?: string | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Event {
  id: number;
  title: string;
  slug: string;
  project?: ('ypsl' | 'ai-driven-sri-lanka' | 'sl-inspire' | 'lets-talk' | 'insl' | 'studpro' | 'y2npro') | null;
  description?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  startDate: string;
  endDate?: string | null;
  startTime?: string | null;
  endTime?: string | null;
  /**
   * Select the primary timezone for this event. Defaults to Sri Lanka (Asia/Colombo).
   */
  timezone?:
    | (
        | 'Asia/Colombo'
        | 'UTC'
        | 'GMT'
        | 'Africa/Abidjan'
        | 'Africa/Accra'
        | 'Africa/Addis_Ababa'
        | 'Africa/Algiers'
        | 'Africa/Asmera'
        | 'Africa/Bamako'
        | 'Africa/Bangui'
        | 'Africa/Banjul'
        | 'Africa/Bissau'
        | 'Africa/Blantyre'
        | 'Africa/Brazzaville'
        | 'Africa/Bujumbura'
        | 'Africa/Cairo'
        | 'Africa/Casablanca'
        | 'Africa/Ceuta'
        | 'Africa/Conakry'
        | 'Africa/Dakar'
        | 'Africa/Dar_es_Salaam'
        | 'Africa/Djibouti'
        | 'Africa/Douala'
        | 'Africa/El_Aaiun'
        | 'Africa/Freetown'
        | 'Africa/Gaborone'
        | 'Africa/Harare'
        | 'Africa/Johannesburg'
        | 'Africa/Juba'
        | 'Africa/Kampala'
        | 'Africa/Khartoum'
        | 'Africa/Kigali'
        | 'Africa/Kinshasa'
        | 'Africa/Lagos'
        | 'Africa/Libreville'
        | 'Africa/Lome'
        | 'Africa/Luanda'
        | 'Africa/Lubumbashi'
        | 'Africa/Lusaka'
        | 'Africa/Malabo'
        | 'Africa/Maputo'
        | 'Africa/Maseru'
        | 'Africa/Mbabane'
        | 'Africa/Mogadishu'
        | 'Africa/Monrovia'
        | 'Africa/Nairobi'
        | 'Africa/Ndjamena'
        | 'Africa/Niamey'
        | 'Africa/Nouakchott'
        | 'Africa/Ouagadougou'
        | 'Africa/Porto-Novo'
        | 'Africa/Sao_Tome'
        | 'Africa/Tripoli'
        | 'Africa/Tunis'
        | 'Africa/Windhoek'
        | 'America/Adak'
        | 'America/Anchorage'
        | 'America/Anguilla'
        | 'America/Antigua'
        | 'America/Araguaina'
        | 'America/Argentina/La_Rioja'
        | 'America/Argentina/Rio_Gallegos'
        | 'America/Argentina/Salta'
        | 'America/Argentina/San_Juan'
        | 'America/Argentina/San_Luis'
        | 'America/Argentina/Tucuman'
        | 'America/Argentina/Ushuaia'
        | 'America/Aruba'
        | 'America/Asuncion'
        | 'America/Bahia'
        | 'America/Bahia_Banderas'
        | 'America/Barbados'
        | 'America/Belem'
        | 'America/Belize'
        | 'America/Blanc-Sablon'
        | 'America/Boa_Vista'
        | 'America/Bogota'
        | 'America/Boise'
        | 'America/Buenos_Aires'
        | 'America/Cambridge_Bay'
        | 'America/Campo_Grande'
        | 'America/Cancun'
        | 'America/Caracas'
        | 'America/Catamarca'
        | 'America/Cayenne'
        | 'America/Cayman'
        | 'America/Chicago'
        | 'America/Chihuahua'
        | 'America/Ciudad_Juarez'
        | 'America/Coral_Harbour'
        | 'America/Cordoba'
        | 'America/Costa_Rica'
        | 'America/Coyhaique'
        | 'America/Creston'
        | 'America/Cuiaba'
        | 'America/Curacao'
        | 'America/Danmarkshavn'
        | 'America/Dawson'
        | 'America/Dawson_Creek'
        | 'America/Denver'
        | 'America/Detroit'
        | 'America/Dominica'
        | 'America/Edmonton'
        | 'America/Eirunepe'
        | 'America/El_Salvador'
        | 'America/Fort_Nelson'
        | 'America/Fortaleza'
        | 'America/Glace_Bay'
        | 'America/Godthab'
        | 'America/Goose_Bay'
        | 'America/Grand_Turk'
        | 'America/Grenada'
        | 'America/Guadeloupe'
        | 'America/Guatemala'
        | 'America/Guayaquil'
        | 'America/Guyana'
        | 'America/Halifax'
        | 'America/Havana'
        | 'America/Hermosillo'
        | 'Asia/Kolkata'
        | 'Asia/Katmandu'
        | 'Asia/Seoul'
        | 'Asia/Singapore'
        | 'Asia/Taipei'
        | 'Asia/Tokyo'
        | 'Europe/Amsterdam'
        | 'Europe/Berlin'
        | 'Europe/London'
        | 'Europe/Paris'
        | 'Europe/Rome'
        | 'Pacific/Auckland'
        | 'Pacific/Honolulu'
      )
    | null;
  eventType: 'physical' | 'online' | 'hybrid';
  venueLocation?: string | null;
  onlinePlatform?: ('google-meet' | 'zoom' | 'other') | null;
  organizers?: (number | Organizer)[] | null;
  image: number | Media;
  registrationUrl?: string | null;
  hashtags?: string | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "organizers".
 */
export interface Organizer {
  id: number;
  name: string;
  logo: number | Media;
  website?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "authors".
 */
export interface Author {
  id: number;
  name: string;
  project?: ('ypsl' | 'ai-driven-sri-lanka' | 'sl-inspire' | 'lets-talk' | 'insl' | 'studpro' | 'y2npro') | null;
  role?: string | null;
  photo?: (number | null) | Media;
  bio?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "articles".
 */
export interface Article {
  id: number;
  title: string;
  slug: string;
  project?: ('ypsl' | 'ai-driven-sri-lanka' | 'sl-inspire' | 'lets-talk' | 'insl' | 'studpro' | 'y2npro') | null;
  content: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  author: number | Author;
  featuredImage: number | Media;
  publishDate?: string | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}